Administrador Proyecto DELETE USERS

+++ PHP/EliminarDocente.php
@@ -10,15 +10,17 @@
     }
 
     
+    $id = null;
     $correo = null;
-	if (!empty($_GET['id'])) {
+	if (!empty($_GET['id']) && !empty($_GET['correo'])) {
 		
-        $correo = $_REQUEST['id'];
+        $correo = $_GET['correo'];
+        $id = $_GET['id'];
         $pdo = Database::connect();
         $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 		$sql = "DELETE FROM PROYECTO_DOCENTE WHERE p_id = ? AND co_correo";
 		$q = $pdo->prepare($sql);
-		$q->execute(array($_SESSION['id'],$correo));
+		$q->execute(array($id,$correo));
 		Database::disconnect();
 		header("Location: AdministradorProyecto.php");
 		exit();
